hunting for warnings

+++ progs/cicada_learn_kbest.cpp
@@ -1870,7 +1870,7 @@ struct OptimizeXBLEU
 	    throw std::runtime_error("no bleu statistics?");
 	    
 	  // collect scaled bleu stats
-	  for (int n = 1; n <= order; ++ n) {
+	  for (size_t n = 1; n <= static_cast<size_t>(order); ++ n) {
 	    if (n - 1 < bleu->ngrams_reference.size())
 	      hypo[n] += prob * bleu->ngrams_reference[n - 1];
 	    if (n - 1 < bleu->ngrams_hypothesis.size())
@@ -1929,7 +1929,7 @@ struct OptimizeXBLEU
 	    const weight_type value(fiter->second * scale);
 	      
 	    // bleu statistics
-	    for (int n = 1; n <= order; ++ n) {
+	    for (size_t n = 1; n <= static_cast<size_t>(order); ++ n) {
 	      if (n - 1 < bleu->ngrams_reference.size())
 		gradients_hypo[n][fiter->first] += value * prob * bleu->ngrams_reference[n - 1];
 		
@@ -1946,7 +1946,7 @@ struct OptimizeXBLEU
 	    
 	  const weight_type value_scale(margin);
 	    
-	  for (int n = 1; n <= order; ++ n) {
+	  for (size_t n = 1; n <= static_cast<size_t>(order); ++ n) {
 	    if (n - 1 < bleu->ngrams_reference.size())
 	      gradients_hypo[n][feature_scale] += value_scale * prob * bleu->ngrams_reference[n - 1];
 	      
@@ -2432,7 +2432,7 @@ struct OptimizeLBFGS
     for (int i = 0; i < threads; ++ i)
       workers.add_thread(new boost::thread(boost::ref(tasks[i])));
     
-    for (int id = 0; id != instances; ++ id)
+    for (int id = 0; id != static_cast<int>(instances); ++ id)
       queue.push(id);
     
     // collect all the objective and gradients...
@@ -2585,7 +2585,7 @@ struct EnvelopeTask
       lines.clear();
       
       for (size_t i = 0; i != kbest_map.size(); ++ i)
-	if (kbest_map[i] == seg)
+	if (kbest_map[i] == static_cast<size_t>(seg))
 	  envelopes(kbests[i].begin(), kbests[i].end(), std::back_inserter(lines));
 
       envelopes(lines);
